refactor: common
- move common functionality into common::
